requirements.txt
Flask==3.0.3
gunicorn==21.2.0

pymongo==4.6.3
certifi==2024.7.4
python-dotenv==1.0.1

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

requests==2.32.3
